Addressed the comments in the code review. Fixed `CSIIndex.getStartOfLastLinearBin` method. Fixed read methods in `CompressedIndexFileBuffer` class. Changed `seek` and `position` methods to work with `long` offsets. Simplified the constructor hierarchy in `AbstractBAMFileIndex`. Added more unit tests in `BAMFileReaderTest`. Added javadoc and annotations.

+++ src/main/java/htsjdk/samtools/IndexFileBuffer.java
@@ -2,12 +2,16 @@ package htsjdk.samtools;
 
 import java.io.Closeable;
 
+/**
+ * Interface for managing index files and streams. Has to be implemented
+ * by all classes that directly access index data.
+ */
 interface IndexFileBuffer extends Closeable {
     void readBytes(final byte[] bytes);
     int readInteger();
     long readLong();
     void skipBytes(final int count);
-    void seek(final int position);
-    int position();
+    void seek(final long position);
+    long position();
     void close();
 }
